Swap provider and consumer, as I wrote them wrong earlier
