+++ contracts/mesh-consumer/src/error.rs
@@ -14,6 +14,18 @@ pub enum ContractError {
     #[error("Unauthorized")]
     Unauthorized {},
 
+    #[error("Contract already has a bound channel: {0}")]
+    ChannelExists(String),
+
+    #[error("Unauthorized counterparty chain, awaiting connection '{0}'")]
+    WrongConnection(String),
+
+    #[error("Unauthorized counterparty port, awaiting port '{0}'")]
+    WrongPort(String),
+
+    #[error("Refuse to respond on unregistered channel '{0}'")]
+    UnknownChannel(String),
+
     #[error("Custom Error val: {val:?}")]
     CustomError { val: String },
     // Add any other custom errors you like here.
