+++ contracts/mesh-consumer/src/ibc.rs
@@ -10,6 +10,7 @@ use cosmwasm_std::{
 use mesh_ibc::{check_order, check_version};
 
 use crate::error::ContractError;
+use crate::state::{CHANNEL, CONFIG};
 
 // TODO: make configurable?
 /// packets live one hour
@@ -18,10 +19,16 @@ pub const PACKET_LIFETIME: u64 = 60 * 60;
 #[cfg_attr(not(feature = "library"), entry_point)]
 /// enforces ordering and versioning constraints
 pub fn ibc_channel_open(
-    _deps: DepsMut,
+    deps: DepsMut,
     _env: Env,
     msg: IbcChannelOpenMsg,
 ) -> Result<Option<Ibc3ChannelOpenResponse>, ContractError> {
+    // ensure we have no other channels currently
+    if let Some(chan) = CHANNEL.may_load(deps.storage)? {
+        return Err(ContractError::ChannelExists(chan));
+    }
+
+    // check the handshake order/version is correct
     let channel = msg.channel();
     check_order(&channel.order)?;
     check_version(&channel.version)?;
@@ -29,33 +36,56 @@ pub fn ibc_channel_open(
         check_version(counter_version)?;
     }
 
-    // TODO: verify this is the authorized connection / port
+    // ensure the remote connection / port is authorized
+    let cfg = CONFIG.load(deps.storage)?;
+    if cfg.provider.connection_id != channel.connection_id {
+        return Err(ContractError::WrongConnection(cfg.provider.connection_id));
+    }
+
+    if cfg.provider.port_id != channel.counterparty_endpoint.port_id {
+        return Err(ContractError::WrongPort(cfg.provider.port_id));
+    }
 
     Ok(None)
 }
 
 #[cfg_attr(not(feature = "library"), entry_point)]
 pub fn ibc_channel_connect(
-    _deps: DepsMut,
+    deps: DepsMut,
     _env: Env,
     msg: IbcChannelConnectMsg,
-) -> StdResult<IbcBasicResponse> {
+) -> Result<IbcBasicResponse, ContractError> {
     let channel = msg.channel();
-    let _channel_id = &channel.endpoint.channel_id;
+    let channel_id = &channel.endpoint.channel_id;
 
-    // TODO: add logic here
+    // save the channel id for future use
+    match CHANNEL.may_load(deps.storage)? {
+        Some(chan) => Err(ContractError::ChannelExists(chan))?,
+        None => CHANNEL.save(deps.storage, channel_id)?,
+    };
 
     Ok(IbcBasicResponse::new())
 }
 
 #[cfg_attr(not(feature = "library"), entry_point)]
 pub fn ibc_channel_close(
-    _deps: DepsMut,
+    deps: DepsMut,
     _env: Env,
-    _msg: IbcChannelCloseMsg,
-) -> StdResult<IbcBasicResponse> {
-    // TODO
-    unimplemented!();
+    msg: IbcChannelCloseMsg,
+) -> Result<IbcBasicResponse, ContractError> {
+    let channel = msg.channel();
+    let channel_id = &channel.endpoint.channel_id;
+
+    // let's ensure this is really closed by same channel we previously connected (paranoia?)
+    // then delete from store
+    let channel = CHANNEL.load(deps.storage)?;
+    if &channel == channel_id {
+        CHANNEL.remove(deps.storage);
+    } else {
+        return Err(ContractError::UnknownChannel(channel_id.clone()));
+    }
+
+    Ok(IbcBasicResponse::new())
 }
 
 #[cfg_attr(not(feature = "library"), entry_point)]
