+++ contracts/mesh-consumer/src/state.rs
@@ -4,13 +4,13 @@ use serde::{Deserialize, Serialize};
 use cosmwasm_std::Addr;
 use cw_storage_plus::Item;
 
-use crate::msg::ConsumerInfo;
+use crate::msg::ProviderInfo;
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 pub struct Config {
     pub count: i32,
     pub owner: Addr,
-    pub consumer: ConsumerInfo,
+    pub provider: ProviderInfo,
 }
 
 pub const CONFIG: Item<Config> = Item::new("config");
