+++ contracts/mesh-provider/src/contract.rs
@@ -21,7 +21,7 @@ pub fn instantiate(
     let state = Config {
         count: msg.count,
         owner: info.sender.clone(),
-        provider: msg.provider,
+        consumer: msg.consumer,
     };
     set_contract_version(deps.storage, CONTRACT_NAME, CONTRACT_VERSION)?;
     CONFIG.save(deps.storage, &state)?;
@@ -88,13 +88,12 @@ pub mod query {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::msg::ProviderInfo;
+    use crate::msg::ConsumerInfo;
     use cosmwasm_std::testing::{mock_dependencies, mock_env, mock_info};
     use cosmwasm_std::{coins, from_binary};
 
-    fn provider_info() -> ProviderInfo {
-        ProviderInfo {
-            port_id: "port-1".to_string(),
+    fn consumer_info() -> ConsumerInfo {
+        ConsumerInfo {
             connection_id: "conn-2".to_string(),
         }
     }
@@ -105,7 +104,7 @@ mod tests {
 
         let msg = InstantiateMsg {
             count: 17,
-            provider: provider_info(),
+            consumer: consumer_info(),
         };
         let info = mock_info("creator", &coins(1000, "earth"));
 
@@ -125,7 +124,7 @@ mod tests {
 
         let msg = InstantiateMsg {
             count: 17,
-            provider: provider_info(),
+            consumer: consumer_info(),
         };
         let info = mock_info("creator", &coins(2, "token"));
         let _res = instantiate(deps.as_mut(), mock_env(), info, msg).unwrap();
@@ -147,7 +146,7 @@ mod tests {
 
         let msg = InstantiateMsg {
             count: 17,
-            provider: provider_info(),
+            consumer: consumer_info(),
         };
         let info = mock_info("creator", &coins(2, "token"));
         let _res = instantiate(deps.as_mut(), mock_env(), info, msg).unwrap();
