+++ contracts/mesh-provider/src/state.rs
@@ -4,13 +4,13 @@ use serde::{Deserialize, Serialize};
 use cosmwasm_std::Addr;
 use cw_storage_plus::Item;
 
-use crate::msg::ProviderInfo;
+use crate::msg::ConsumerInfo;
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 pub struct Config {
     pub count: i32,
     pub owner: Addr,
-    pub provider: ProviderInfo,
+    pub consumer: ConsumerInfo,
 }
 
 pub const CONFIG: Item<Config> = Item::new("config");
